UC-157: Account Authorization

+++ lib_auth/src/main/java/com/ucm/lib/dao/UserDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDAO extends JpaRepository<User, Long> {
+public interface UserDAO extends JpaRepository<User, Integer> {
     User findByUsername(String username);
 
     Boolean existsByUsername(String username);
